fix return size of get_amount_recovery_result_size

+++ src/blsct/external_api/blsct.cpp
@@ -354,7 +354,7 @@ void delete_amount_recovery_req_vec(void* vp_amt_recovery_req_vec) {
   }
 
   // functions to retrieve attrs of amount recovery result
-int16_t get_amount_recovery_result_size(
+size_t get_amount_recovery_result_size(
     void* vp_amt_recovery_res_vec
 ) {
     if (vp_amt_recovery_res_vec == nullptr) {
